Fibonacci number

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,23 +9,23 @@
 int main(void)
 {
 	int i;
-	int fab[50];
+	long int fab[50];
 
 	fab[0] = 1;
 	fab[1] = 2;
 
-	printf("%d, %d, ", fab[0], fab[1]);
+	printf("%ld, %ld, ", fab[0], fab[1]);
 
 	for (i = 2; i < 50; i++)
 	{
 		fab[i] = fab[i - 1] + fab[i - 2];
 		if (i == 49)
 		{
-			printf("%d\n", fab[i]);
+			printf("%ld\n", fab[i]);
 		}
 		else
 		{
-			printf("%d, ", fab[i]);
+			printf("%ld, ", fab[i]);
 		}
 	}
 	return (0);
